fixed hw

+++ 07-Primera Entrega/controllers/carrito.js
@@ -1,7 +1,19 @@
 const Contenedor = require("./../contenedor");
 const path = require("path");
+const uuid = require("uuid");
+
+
+let administrador = true;
 
 async function getAllCartProducts(req,res){
+
+    if(!administrador){
+        return res.status(400).json({
+            success: false,
+            message: "You don't have permissions"
+        })
+    }
+
     const contenedor = new Contenedor(path.join(__dirname,"..","data","carrito.txt"));
     const carrito = await contenedor.getById(req.params.id);
 
@@ -14,6 +26,16 @@ async function getAllCartProducts(req,res){
 }
 
 async function getCart(req,res){
+
+    if(!administrador){
+        return res.status(400).json({
+            success: false,
+            message: "You don't have permissions"
+        })
+    }
+
+
+
     const contenedor = new Contenedor(path.join(__dirname,"..","data","carrito.txt"));
     const id = req.params.id;
     const carrito = await contenedor.getById(id);
@@ -29,6 +51,17 @@ async function getCart(req,res){
 }
 
 async function createCart(req,res){
+
+
+    if(!administrador){
+        return res.status(400).json({
+            success: false,
+            message: "You don't have permissions"
+        })
+    }
+
+
+
     const contenedor = new Contenedor(path.join(__dirname,"..","data","carrito.txt"));
     if(!req.body){
         return res.status(400).json({
@@ -44,9 +77,25 @@ async function createCart(req,res){
 }
 
 async function createCartProduct(req,res){
+
+    if(!administrador){
+        return res.status(400).json({
+            success: false,
+            message: "You don't have permissions"
+        })
+    }
+
+    const producto = req.body;
+
+    producto.id = uuid.v4();
+    producto.timestamp = Date.now();
+
+
+
+
     const contenedor = new Contenedor(path.join(__dirname,"..","data","carrito.txt"));
     let cart = await contenedor.getById(req.params.id);
-    cart.push(req.body);
+    cart?.productos.push(producto);
     const cartId = await contenedor.updateById(req.params.id,cart);
     if(!cartId){
         return res.status(400).json({
@@ -60,6 +109,16 @@ async function createCartProduct(req,res){
 }
 
 async function deleteCart(req,res){
+
+
+    if(!administrador){
+        return res.status(400).json({
+            success: false,
+            message: "You don't have permissions"
+        })
+    }
+
+
     const contenedor = new Contenedor(path.join(__dirname,"..","data","carrito.txt"));
     const cartId = await contenedor.deleteById(req.params.id);
     if(!cartId){
@@ -74,10 +133,21 @@ async function deleteCart(req,res){
 }
 
 async function deleteCartProduct(req,res) {
+
+    if(!administrador){
+        return res.status(400).json({
+            success: false,
+            message: "You don't have permissions"
+        })
+    }
+
+
+
     const contenedor = new Contenedor(path.join(__dirname,"..","data","carrito.txt"));
     let cart = await contenedor.getById(req.params.id);
-    cart = cart.filter((product) => product.id != req.params["id_prod"])
-    await contenedor.updateById(cart,req.params.id);
+    console.log("El carrito es: " ,cart);
+    cart.productos = cart.productos.filter((product) => product?.id !== req.params["id_prod"])
+    await contenedor.updateById(req.params.id,cart);
     return res.status(200).json({
         success: true,
         data: cart.id
